Avoid using confusing "ITest" name for tests that are not QuarkusIntegrationTests
