+++ integration-tests/kubernetes-client-devservices/src/test/java/io/quarkus/kubernetes/client/devservices/it/DevServicesKubernetesTest.java
@@ -13,7 +13,7 @@ import io.quarkus.test.junit.TestProfile;
 
 @QuarkusTest
 @TestProfile(DevServiceKubernetes.class)
-public class DevServicesKubernetesITest {
+public class DevServicesKubernetesTest {
 
     @Inject
     KubernetesClient kubernetesClient;
