+++ integration-tests/redis-devservices/src/test/java/io/quarkus/redis/devservices/it/DevServicesRedisRandomPortTest.java
@@ -1,7 +1,5 @@
 package io.quarkus.redis.devservices.it;
 
-import static org.junit.jupiter.api.Assertions.fail;
-
 import jakarta.inject.Inject;
 
 import org.junit.jupiter.api.Assertions;
@@ -10,13 +8,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import io.quarkus.redis.datasource.RedisDataSource;
-import io.quarkus.redis.devservices.it.profiles.DevServicesNonUniquePortProfile;
+import io.quarkus.redis.devservices.it.profiles.DevServicesRandomPortProfile;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
 
 @QuarkusTest
-@TestProfile(DevServicesNonUniquePortProfile.class)
-public class DevServicesRedisNonUniquePortITest {
+@TestProfile(DevServicesRandomPortProfile.class)
+public class DevServicesRedisRandomPortTest {
 
     @Inject
     RedisDataSource redisClient;
@@ -30,16 +28,6 @@ public class DevServicesRedisNonUniquePortITest {
     @DisplayName("given redis container must communicate with it and return value by key")
     public void shouldReturnAllKeys() {
         Assertions.assertEquals("anyvalue", redisClient.value(String.class).get("anykey").toString());
-    }
-
-    @Test
-    public void shouldHaveStack() {
-        // The entertainingly named lolwut command gives a version number but doesn't say if stack is available, so try using some of the commands
-        try {
-            redisClient.bloom(String.class).bfadd("key", "whatever");
-        } catch (Exception e) {
-            fail("Redis stack should be available on the connected back end (underlying error was  " + e + ")");
-        }
     }
 
 }
